+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Add `object-list-view-cell` component.
 * Add locales for title attribute in header of `object-list-view` component.
 
+### Fixed
+* Displaying sorting indicator in `object-list-view` component.
+
 ## [0.7.0-beta.10] - 2016-10-06
 ### Changed
 * Updated dependency on `ember-flexberry-data` addon to v0.6.2-beta.13 version.